requirements.txt
firecrawl-py
openai
SQLAlchemy
python-dotenv